Moved UseBlockCallback into its own PlayerUseSignHandler.kt class Reformated some classes

+++ src/main/java/fr/unreal852/quantum/utils/ParseUtils.kt
@@ -2,11 +2,11 @@ package fr.unreal852.quantum.utils
 
 object ParseUtils {
     @JvmStatic
-    fun tryParseLong(value: String, defaultVal: Int): Long {
+    fun tryParseLong(value: String, defaultVal: Long): Long {
         return try {
             value.toLong()
         } catch (e: NumberFormatException) {
-            defaultVal.toLong()
+            defaultVal
         }
     }
 }
